cloud-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
firebase-admin==6.2.0
pydantic==2.5.0
python-multipart==0.0.6
httpx==0.25.2
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
pydantic_settings
pydantic[email]
google-generativeai
rich